Move the extract_dedupe_value and create_occurrence method to top level detector handler.
