+++ src/sentry/workflow_engine/handlers/detector/__init__.py
@@ -1,4 +1,6 @@
 __all__ = [
+    "DataPacketEvaluationType",
+    "DataPacketType",
     "DetectorEvaluationResult",
     "DetectorHandler",
     "DetectorOccurrence",
@@ -6,5 +8,11 @@ __all__ = [
     "StatefulGroupingDetectorHandler",
 ]
 
-from .base import DetectorEvaluationResult, DetectorHandler, DetectorOccurrence
+from .base import (
+    DataPacketEvaluationType,
+    DataPacketType,
+    DetectorEvaluationResult,
+    DetectorHandler,
+    DetectorOccurrence,
+)
 from .stateful import DetectorStateData, StatefulGroupingDetectorHandler
